% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elbow.R
\name{plot_elbow}
\alias{plot_elbow}
\title{Plot Elbow Method for KMeans}
\usage{
plot_elbow(data)
}
\arguments{
\item{data}{A numeric matrix or data frame for clustering.}
}
\value{
A ggplot object showing the elbow plot.
}
\description{
Uses within-cluster sum of squares (WSS) to help determine the optimal number of clusters.
}
\examples{
data <- scale(iris[, 1:4])
if (interactive()) {
  plot_elbow(data)
}


}
