% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar.R
\name{plot_radar}
\alias{plot_radar}
\title{Plot Radar Chart for PAM Cluster Centers}
\usage{
plot_radar(data, clusters)
}
\arguments{
\item{data}{A numeric matrix or data frame for clustering.}

\item{clusters}{An integer indicating the number of clusters.}
}
\value{
A ggplot object showing the radar chart of cluster medoids.
}
\description{
Displays the medoids of each PAM cluster using a polar radar chart.
}
\examples{
data <- scale(iris[, 1:4])
if (interactive()) {
  plot_radar(data, clusters = 3)
}


}
