% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Shiny Clustering Web App}
\usage{
run_app()
}
\value{
No return value. This function is called for its side effect (launching the app).
}
\description{
This function launches the Shiny web application located in the \code{inst/app} directory
of the installed package. The application provides an interactive interface for clustering analysis.
}
\examples{
if (interactive()) {
  run_app()
}


}
